requirements.txt
numpy==2.4.2
pandas==2.3.3
streamlit==1.54.0
scikit-learn==1.8.0
xgboost-cpu==3.2.0
scipy==1.17.1